# Go_Server/docs/swagger.yaml
definitions:
...
  define.AddOrderRequest:
    properties:
      food:
        type: string
      num:
        type: integer
      remarks:
        type: string
      user:
        type: string
    type: object
  define.AddRoleRequest:
    properties:
      level:
        type: integer
      name:
        type: string
      remarks:
        type: string
    type: object
...
  define.GetOrderListRequest:
    properties:
      keyword:
        type: string
      page:
        type: integer
      size:
        type: integer
      status:
        type: integer
    type: object
  define.GetRoleListRequest:
    properties:
      keyword:
        type: string
      page:
        type: integer
      size:
        type: integer
      status:
        type: integer
    type: object
...
  define.UpdateRoleRequest:
    properties:
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      remarks:
        type: string
    type: object
...
info:
  contact:
    name: Vulnerabilities_Server
    url: https://github.com/A7cc/Vulnerabilities_Server
  description: 这是一个用Golang写的食谱菜单管理系统靶场，该系统中存在多个安全漏洞，需要我们去探索和发现。\n\n该项目旨在帮助安全研究人员和爱好者了解和掌握关于Golang系统的渗透测试和代码审计知识。
  title: 食谱菜单管理系统靶场
  version: "1.0"
paths:
...
  /order/add:
    post:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 添加订单信息
        in: body
        name: AddOrderRequest
        required: true
        schema:
          $ref: '#/definitions/define.AddOrderRequest'
      responses: {}
      summary: 新增订单信息
      tags:
      - 鉴权接口-订单相关方法
  /order/delete/{id}:
    delete:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除订单ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 删除订单信息
      tags:
      - 鉴权接口-订单相关方法
  /order/detail:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取订单ID
        in: query
        name: id
        required: true
        type: string
      responses: {}
      summary: 根据ID获取订单信息
      tags:
      - 鉴权接口-订单相关方法
  /order/get:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取订单查询参数
        in: body
        name: GetOrderListRequest
        required: true
        schema:
          $ref: '#/definitions/define.GetOrderListRequest'
      responses: {}
      summary: 获取订单列表
      tags:
      - 鉴权接口-订单相关方法
  /role/add:
    post:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 添加角色信息
        in: body
        name: AddRoleRequest
        required: true
        schema:
          $ref: '#/definitions/define.AddRoleRequest'
      responses: {}
      summary: 新增角色信息
      tags:
      - 鉴权接口-角色相关方法
  /role/delete/{id}:
    delete:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除角色ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 删除角色信息
      tags:
      - 鉴权接口-角色相关方法
  /role/detail:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取角色ID
        in: query
        name: id
        required: true
        type: integer
      responses: {}
      summary: 根据ID获取角色信息
      tags:
      - 鉴权接口-角色相关方法
  /role/get:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取角色查询参数
        in: body
        name: GetRoleListRequest
        required: true
        schema:
          $ref: '#/definitions/define.GetRoleListRequest'
      responses: {}
      summary: 获取角色列表信息
      tags:
      - 鉴权接口-角色相关方法
  /role/update:
    put:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新角色信息参数
        in: body
        name: UpdateRoleRequest
        required: true
        schema:
          $ref: '#/definitions/define.UpdateRoleRequest'
      responses: {}
      summary: 修改角色信息
      tags:
      - 鉴权接口-角色相关方法
  /settings/backupsdb:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 备份数据库
      tags:
      - 鉴权接口-系统设置相关方法
  /settings/deletedb:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除数据库备份文件
        in: body
        name: dbfile
        required: true
        schema:
          type: string
      responses: {}
      summary: 删除数据库
      tags:
      - 鉴权接口-系统设置相关方法
  /settings/downdb:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 下载数据库备份文件
        in: body
        name: dbfile
        required: true
        schema:
          type: string
      responses: {}
      summary: 数据库下载
      tags:
      - 鉴权接口-系统设置相关方法
  /settings/getdb:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 数据库备份目录
        in: body
        name: dir
        required: true
        schema:
          type: string
      responses: {}
      summary: 获取数据库列表
      tags:
      - 鉴权接口-系统设置相关方法
...
swagger: "2.0"